# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Reisetagebuch API Draft",
        "description": "This is a draft version of the Reisetagebuch API documentation. The API is still under development and may change in future releases. Use at your own risk!",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "0.1.0-draft",
        "x-logo": {
            "url": "https://raw.githubusercontent.com/HerrLevin/reisetagebuch/refs/heads/main/public/assets/logo.webp"
        }
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/timeline": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get timeline posts",
                "description": "Returns paginated posts for the authenticated user timeline",
                "operationId": "timeline",
                "parameters": [
                    {
                        "name": "cursor",
                        "in": "query",
                        "description": "Pagination cursor",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostPaginationDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/users/{userId}/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get posts for a specific user",
                "description": "Returns paginated posts for a specific user",
                "operationId": "postsForUser",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "cursor",
                        "in": "query",
                        "description": "Pagination cursor",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostPaginationDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/posts/{id}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get post by ID",
                "description": "Returns post data",
                "operationId": "showPost",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Post id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/BasePost"
                                        },
                                        {
                                            "$ref": "#/components/schemas/TransportPost"
                                        },
                                        {
                                            "$ref": "#/components/schemas/LocationPost"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Posts"
                ],
                "summary": "Delete post",
                "description": "Delete a post by id",
                "operationId": "deletePost",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Post id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Posts"
                ],
                "summary": "Update post",
                "description": "Update a post (text, transport or location)",
                "operationId": "updatePost",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Post id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "App\\Http\\Requests\\BasePostRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/BasePost"
                                        },
                                        {
                                            "$ref": "#/components/schemas/TransportPost"
                                        },
                                        {
                                            "$ref": "#/components/schemas/LocationPost"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Filter posts",
                "description": "Returns filtered posts",
                "operationId": "filterPosts",
                "parameters": [
                    {
                        "name": "cursor",
                        "in": "query",
                        "description": "Pagination cursor",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "dateFrom",
                        "in": "query",
                        "description": "Filter posts from this date (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "dateTo",
                        "in": "query",
                        "description": "Filter posts to this date (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "visibility",
                        "in": "query",
                        "description": "Filter by visibility (e.g., PUBLIC, PRIVATE)",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/components/schemas/Visibility"
                            }
                        }
                    },
                    {
                        "name": "travelReason",
                        "in": "query",
                        "description": "Filter by travel reason (e.g., WORK, LEISURE)",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/components/schemas/TravelReason"
                            }
                        }
                    },
                    {
                        "name": "tags",
                        "in": "query",
                        "description": "Filter by tags",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FilteredPostPaginationDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/posts/mass-edit": {
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Mass edit posts",
                "description": "Mass edit multiple posts",
                "operationId": "massEditPosts",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "App\\Http\\Requests\\MassEditPostRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/posts/transport": {
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Create transport post",
                "description": "Create a transport post",
                "operationId": "storeTransportPost",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "App\\Http\\Requests\\TransportBasePostCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TransportPost"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "TravelReason": {
                "description": "Reasons for travel associated with transport posts",
                "type": "string",
                "enum": [
                    "commute",
                    "business",
                    "leisure",
                    "crew",
                    "errand",
                    "other"
                ]
            },
...
            "Visibility": {
                "description": "Visibility levels for posts",
                "type": "string",
                "enum": [
                    "public",
                    "private",
                    "unlisted",
                    "only-authenticated"
                ]
            },
...
        },
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
...
    "security": [
        {
            "passport": []
        }
    ]
}